requirements.txt
gitpython
fastapi
uvicorn
torch<2.6.0
torchvision<=0.20
pillow
numpy
opencv-python-headless
pandas
IPython
requests
PyYAML>=5.3.1
psutil
tqdm
matplotlib
seaborn
scipy
albumentations>=1.0.3
pycocotools>=2.0
